backend/requirements.txt
asgiref==3.7.2
Django==3.2.22
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
environ==1.0
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
